chore: move swagger definitions to dedicated module in `auth` app
